+++ test/unit/lib/device.js
@@ -34,7 +34,8 @@ var AGENTS = {
   pcChrome_27: 'Mozilla/5.0 (Windows NT 6.1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/27.0.1453.93 Safari/537.36',
   pcChrome_41: 'Mozilla/5.0 (Windows NT 6.1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/41.0.2228.0 Safari/537.36',
   pcFirefox: 'Mozilla/5.0 (Windows NT x.y; Win64; x64; rv:10.0) Gecko/20100101 Firefox/10.0',
-  pcSafari5_1: 'Mozilla/5.0 (Windows; U; Windows NT 6.1; en-us) AppleWebKit/534.50 (KHTML, like Gecko) Version/5.1 Safari/534.50'
+  pcSafari5_1: 'Mozilla/5.0 (Windows; U; Windows NT 6.1; en-us) AppleWebKit/534.50 (KHTML, like Gecko) Version/5.1 Safari/534.50',
+  samsungBrowser2_1: 'Mozilla/5.0 (Linux; Android 5.0.1; SAMSUNG SPH-L720T Build/LRX22C) AppleWebKit/537.36 (KHTML, like Gecko) SamsungBrowser/2.1 Chrome/34.0.1847.76 Mobile Safari/537.36'
 };
 
 describe('device', function () {
@@ -93,6 +94,26 @@ describe('device', function () {
     it('returns false for non-IE9', function () {
       expect(device.isIE9(AGENTS.ie10)).to.equal(false);
       expect(device.isIE9(AGENTS.ie11)).to.equal(false);
+    });
+  });
+
+  describe('isSamsungBrowser', function () {
+    it('returns true for Samsung Browser', function () {
+      expect(device.isSamsungBrowser(AGENTS.samsungBrowser2_1)).to.equal(true);
+    });
+
+    it('returns false when not Samsung Browser', function () {
+      var key, ua;
+
+      for (key in AGENTS) {
+        if (!AGENTS.hasOwnProperty(key)) {
+          continue;
+        }
+        if (!/samsungBrowser/.test(key)) {
+          ua = AGENTS[key];
+          expect(device.isSamsungBrowser(ua)).to.be.false;
+        }
+      }
     });
   });
 });